refactor: move basic implementations into separate module
