+++ crates/bitwarden-error-macro/src/basic/mod.rs
@@ -1,0 +1,2 @@
+pub mod attribute;
+pub mod derive;
